% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df_text}
\alias{clean_df_text}
\title{Clean a text variable of a df}
\usage{
clean_df_text(
  vec_text,
  recode_words = get_recode_words(),
  stopwords = get_specific_stopwords(),
  multiwords = get_specific_multiwords()
)
}
\arguments{
\item{vec_text}{vector of texts}

\item{recode_words}{names vector of words to recode}

\item{stopwords}{words to remove}

\item{multiwords}{words to collapse}
}
\value{
vector of character
}
\description{
Clean a text variable of a df
}
\examples{
clean_df_text(janeaustenr::austen_books(),"text")
}
