% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{col2hex}
\alias{col2hex}
\title{Convert color to hexadecimal}
\usage{
col2hex(x, alpha = FALSE)
}
\arguments{
\item{x}{name of color}

\item{alpha}{get transparent color}
}
\value{
hex color
}
\description{
Convert color to hexadecimal
}
\examples{
col2hex('red')
}
