% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.R
\name{corpus_to_tokens}
\alias{corpus_to_tokens}
\title{Convert quanteda corpus to tokens}
\usage{
corpus_to_tokens(
  corpus,
  recode_words = get_recode_words(),
  stopwords = get_specific_stopwords(),
  multiwords = get_specific_multiwords()
)
}
\arguments{
\item{corpus}{corpus from quanteda}

\item{recode_words}{names vector of words to recode}

\item{stopwords}{words to remove}

\item{multiwords}{words to collapse}
}
\value{
tokens
}
\description{
Convert quanteda corpus to tokens
}
\examples{
corpus_to_tokens(quanteda::data_corpus_inaugural)
}
