% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq.R
\name{export_to_iramuteq}
\alias{export_to_iramuteq}
\title{Export to iramuteq}
\usage{
export_to_iramuteq(df, meta_cols, text_col, output_file)
}
\arguments{
\item{df}{df to export}

\item{meta_cols}{names of cols to export as meta}

\item{text_col}{name of col to export as text}

\item{output_file}{name of txt file}
}
\value{
nothing (create file)
}
\description{
Export to iramuteq
}
\examples{
df <- janeaustenr::austen_books()[1:10,]
path_txt <- file.path(tempfile(fileext=".txt"))
export_to_iramuteq(df,"book","text",path_txt)
}
