% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{format_emotions}
\alias{format_emotions}
\title{Format gt with emotions}
\usage{
format_emotions(gt)
}
\arguments{
\item{gt}{gt to format : label, color, percent}
}
\value{
gt
}
\description{
Format gt with emotions
}
\examples{
vec_text <- janeaustenr::austen_books()$text
sentiment_scores <- get_nrc_sentiment(vec_text[1:50], lang="english")
df_emotion <- get_dominant_emotion(df_sentiment,1)
format_emotions(gt(summarise_emotions(df_emotion)))
}
