% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{format_polarity}
\alias{format_polarity}
\title{Format gt with polarity}
\usage{
format_polarity(gt)
}
\arguments{
\item{gt}{gt to format : label, color, percent}
}
\value{
gt
}
\description{
Format gt with polarity
}
\examples{
vec_text <- janeaustenr::austen_books()$text
sentiment_scores <- syuzhet::get_nrc_sentiment(vec_text[1:50], lang="english")
df_polarity <- get_dominant_polarity(df_sentiment,1)
format_polarity(gt(summarise_polarity(df_polarity)))
}
