% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get quanteda dictionary with preset terms}
\usage{
get_dictionary()
}
\value{
quanteda dictionary
}
\description{
Get quanteda dictionary with preset terms
}
\examples{
get_dictionary()
}
