% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_dominant_emotion}
\alias{get_dominant_emotion}
\title{Get most frequent emotion of a sentiment score df}
\usage{
get_dominant_emotion(df_sentiment, threshold = 5)
}
\arguments{
\item{df_sentiment}{data.frame from get_nrc_sentiment}

\item{threshold}{minimum frequency to keep emotion}
}
\value{
data.frame
}
\description{
Get most frequent emotion of a sentiment score df
}
\examples{
vec_text <- janeaustenr::austen_books()$text
df_sentiment <- syuzhet::get_nrc_sentiment(vec_text[1:50], lang="english")
df_emotion <- get_dominant_emotion(df_sentiment,2)
dplyr::count(df_emotion,emotion)
}
