% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{get_dominant_polarity}
\alias{get_dominant_polarity}
\title{Get most frequent polarity of a sentiment score df}
\usage{
get_dominant_polarity(df_sentiment, threshold = 5)
}
\arguments{
\item{df_sentiment}{data.frame from get_nrc_sentiment}

\item{threshold}{minimum frequency to keep polarity}
}
\value{
data.frame
}
\description{
Get most frequent polarity of a sentiment score df
}
\examples{
vec_text <- janeaustenr::austen_books()$text
df_sentiment <- get_nrc_sentiment(vec_text[1:50], lang="english")
df_polarity <- get_dominant_polarity(df_sentiment,2)
count(df_polarity,polarity)
}
