% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_specific_multiwords}
\alias{get_specific_multiwords}
\title{Get specific multiwords for the study}
\usage{
get_specific_multiwords()
}
\value{
vector of terms
}
\description{
Get specific multiwords for the study
}
\examples{
get_specific_multiwords()
}
