% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_specific_stopwords}
\alias{get_specific_stopwords}
\title{Specific stopwords for this study}
\usage{
get_specific_stopwords()
}
\value{
vector of terms
}
\description{
Specific stopwords for this study
}
\examples{
get_specific_stopwords()
}
