% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_videos_details}
\alias{get_videos_details}
\title{Get details from youtube videos}
\usage{
get_videos_details(api_key, video_ids)
}
\arguments{
\item{api_key}{youtube api key}

\item{video_ids}{vector of youtube id}
}
\value{
data.frame
}
\description{
Get details from youtube videos
}
