% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_videos_info}
\alias{get_videos_info}
\title{Get YouTube informations of videos from a playlist}
\usage{
get_videos_info(api_key, playlist_id, max_results = 100)
}
\arguments{
\item{api_key}{youtube api key}

\item{playlist_id}{id of the playlist}

\item{max_results}{maximum amount of results}
}
\value{
data.frame
}
\description{
Get YouTube informations of videos from a playlist
}
