% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_videos_stat}
\alias{get_videos_stat}
\title{Get YouTube statistics informations of videos}
\usage{
get_videos_stat(api_key, video_ids)
}
\arguments{
\item{api_key}{youtube api key}

\item{video_ids}{vector of youtube id}
}
\value{
data.frame
}
\description{
Get YouTube statistics informations of videos
}
