% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_videos_text}
\alias{get_videos_text}
\title{Get YouTube subtitles of videos}
\usage{
get_videos_text(
  video_ids,
  path,
  suffix,
  yt_dlp_path = "yt-dlp",
  force_dl = FALSE
)
}
\arguments{
\item{video_ids}{vector of youtube id}

\item{path}{path to save subtitles}

\item{suffix}{suffix to identify the playlist}

\item{yt_dlp_path}{path to yt_dlp}

\item{force_dl}{force the download of subtitles}
}
\value{
data.frame
}
\description{
Get YouTube subtitles of videos
}
