% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitle.R
\name{group_minuted_text}
\alias{group_minuted_text}
\title{Regroup text by minutes}
\usage{
group_minuted_text(df, minutes, video_id = "video_id")
}
\arguments{
\item{df}{minuted data.frame from read_vtt_as_df()}

\item{minutes}{number of minutes to regroup}

\item{video_id}{name of id variable}
}
\value{
data.frame
}
\description{
Regroup text by minutes
}
\examples{
df_text <- read.csv("inst/extdata/df_text_bfm.csv")
head(group_minuted_text(df_text,2))
}
