% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitle.R
\name{hms_to_sec}
\alias{hms_to_sec}
\title{Convert hour:minutes:secondes to secondes}
\usage{
hms_to_sec(x)
}
\arguments{
\item{x}{pattern to convert}
}
\value{
float
}
\description{
Convert hour:minutes:secondes to secondes
}
\examples{
hms_to_sec("00:25:30.389")
}
