% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq.R
\name{import_from_iramuteq}
\alias{import_from_iramuteq}
\title{import df from iramuteq}
\usage{
import_from_iramuteq(file)
}
\arguments{
\item{file}{path to file .txt from IRaMuTeQ}
}
\value{
data.frame
}
\description{
import df from iramuteq
}
\examples{
df <- janeaustenr::austen_books()[1:10,]
path_txt <- file.path(tempfile(fileext=".txt"))
export_to_iramuteq(df,"book","text",path_txt)
import_from_iramuteq(path_txt)
}
