% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lexico_example}
\alias{lexico_example}
\title{Obtain extdata for the lexico package}
\usage{
lexico_example(path = NULL)
}
\arguments{
\item{path}{name of file}
}
\value{
link of path or (if path is null) list of paths
}
\description{
Obtain extdata for the lexico package
}
\examples{
lexico_example()
lexico_example("df_info_bfm.csv")
}
