% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{parse_yt_duration}
\alias{parse_yt_duration}
\title{Get duration of a youtube video}
\usage{
parse_yt_duration(duration)
}
\arguments{
\item{duration}{duration in youtube format}
}
\value{
number
}
\description{
Get duration of a youtube video
}
\examples{
parse_yt_duration("PT25M31S")
}
