% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq.R
\name{read_iramuteq_class}
\alias{read_iramuteq_class}
\title{Read report from iramuteq to extract words of each class}
\usage{
read_iramuteq_class(file, classe = 1)
}
\arguments{
\item{file}{path to report file .txt from IRaMuTeQ}

\item{classe}{index of class to extract}
}
\value{
data.frame
}
\description{
Read report from iramuteq to extract words of each class
}
\examples{
path_txt <- lexico_example("RAPPORT.txt")
head(read_iramuteq_class(path_txt,1))
}
