% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitle.R
\name{read_vtt_as_df}
\alias{read_vtt_as_df}
\title{Convert subtitles file to minuted data.frame}
\usage{
read_vtt_as_df(vtt_file)
}
\arguments{
\item{vtt_file}{name of a subtitles file}
}
\value{
data.frame
}
\description{
Convert subtitles file to minuted data.frame
}
\examples{
path_vtt <- lexico_example("6IOUEJN6GRI.fr.vtt")
read_vtt_as_df(path_vtt)
}
