% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{read_vtt_as_text}
\alias{read_vtt_as_text}
\title{Convert subtitles file to text}
\usage{
read_vtt_as_text(vtt_file)
}
\arguments{
\item{vtt_file}{name of a subtitles file}
}
\value{
data.frame
}
\description{
Convert subtitles file to text
}
