% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{remove_apostrophe}
\alias{remove_apostrophe}
\title{Correct apostrophe}
\usage{
remove_apostrophe(
  text,
  patterns = c("l", "d", "qu", "j", "s", "m", "n", "c", "t")
)
}
\arguments{
\item{text}{vector of strings to correct}

\item{patterns}{patterns}
}
\description{
Correct apostrophe
}
