% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{run_complete_extraction}
\alias{run_complete_extraction}
\title{Run a complete extraction for a playlist id}
\usage{
run_complete_extraction(api_key, path, suffix, playlist_id, max_videos)
}
\arguments{
\item{api_key}{youtube api key}

\item{path}{path to save subtitles}

\item{suffix}{suffix to identify the playlist}

\item{playlist_id}{id of the playlist}

\item{max_videos}{maximum amont of videos to extract}
}
\description{
Run a complete extraction for a playlist id
}
