% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{summarise_emotions}
\alias{summarise_emotions}
\title{Summarise emotions of a data.frame}
\usage{
summarise_emotions(df)
}
\arguments{
\item{df}{data.frame to summarise}
}
\value{
data.frame
}
\description{
Summarise emotions of a data.frame
}
\examples{
vec_text <- janeaustenr::austen_books()$text
df_sentiment <- syuzhet::get_nrc_sentiment(vec_text[1:50], lang="english")
df_emotion <- get_dominant_emotion(df_sentiment,1)
summarise_emotions(df_emotion)
}
