% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syuzhet.R
\name{summarise_polarity}
\alias{summarise_polarity}
\title{Summarise polarity of a data.frame}
\usage{
summarise_polarity(df)
}
\arguments{
\item{df}{data.frame to summarise}
}
\value{
data.frame
}
\description{
Summarise polarity of a data.frame
}
\examples{
vec_text <- janeaustenr::austen_books()$text
df_sentiment <- syuzhet::get_nrc_sentiment(vec_text[1:50], lang="english")
df_polarity <- get_dominant_polarity(df_sentiment,1)
summarise_polarity(df_polarity)
}
