% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{treemap_double_classe}
\alias{treemap_double_classe}
\title{Plot a treemap of thema and classe of segments}
\usage{
treemap_double_classe(
  df,
  id_theme,
  theme,
  classe,
  threshold = 30,
  title = "Treemap",
  footer = "",
  palette = NA
)
}
\arguments{
\item{df}{data.frame with theme and classe}

\item{id_theme}{id of theme (to order it)}

\item{theme}{name of theme variable}

\item{classe}{name of classe variable}

\item{threshold}{minimum frequent of classe to show}

\item{title}{title of the treemap}

\item{footer}{(optional) footer of the treemap}

\item{palette}{(optional) palette to color theme}
}
\value{
plot
}
\description{
Plot a treemap of thema and classe of segments
}
\examples{
df_segment <- read.csv(lexico_example("df_segment.csv"))
treemap_double_classe(df_segment,"id_classe","classe","classe_local",0)
}
