% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{vtt_files_to_df}
\alias{vtt_files_to_df}
\title{Convert vtt files to minuted data.frame}
\usage{
vtt_files_to_df(path, suffix)
}
\arguments{
\item{path}{path of vtt files}

\item{suffix}{suffix to identify the playlist}
}
\value{
data.frame
}
\description{
Convert vtt files to minuted data.frame
}
