% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{download_subtitles}
\alias{download_subtitles}
\title{Download subtitles from a youtube video}
\usage{
download_subtitles(video_id, out_dir, yt_dlp = "yt-dlp", force_dl = FALSE)
}
\arguments{
\item{video_id}{video id}

\item{out_dir}{directory to save file}

\item{yt_dlp}{name of program to launch}
}
\description{
Download subtitles from a youtube video
}
