% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq.R
\name{import_from_iramuteq}
\alias{import_from_iramuteq}
\title{import df from iramuteq}
\usage{
import_from_iramuteq(file)
}
\arguments{
\item{file}{}
}
\value{
data.frame
}
\description{
import df from iramuteq
}
\examples{
df <- janeaustenr::austen_books()
export_to_iramuteq(df,"book","text",tempfile())
import_from_iramuteq(tempfile())
}
