% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitle.R
\name{read_vtt_as_text}
\alias{read_vtt_as_text}
\title{Convert subtitles file to text (OLD)}
\usage{
read_vtt_as_text(vtt_file)
}
\arguments{
\item{vtt_file}{name of a subtitles file}
}
\value{
data.frame
}
\description{
Convert subtitles file to text (OLD)
}
\examples{
path_vtt <- lexico_example("6IOUEJN6GRI.fr.vtt")
read_vtt_as_text(path_vtt)

}
